refactor(data): name every path icon after its path
An icon belongs to the path, not to one of its two martial arts — both arts
of a class import the same file, and `ClassPicker` shows it for the class. The
two shipped icons were named for the sword and the blade that happened to
carry them, which read as that art's icon rather than the path's.

Renamed to match the six staged last commit, so the folder is now one file
per path with no exceptions.

Checked with a build: a renamed asset that some import still points at is a
bundler error rather than a test failure.
