+++ src/data/martialArts/icons/bellstrike-umbra.png
