+++ src/data/martialArts/icons/stonesplit-strength.png
