+++ src/data/martialArts/phalanxbaneBlade.ts
@@ -1,6 +1,6 @@
 import { defineMartialArt } from "../../definitions/martialArts/martialArtDef"
 import { MARTIAL_ART_ID } from "./ids"
-import icon from "./icons/snowparting-blade.png"
+import icon from "./icons/stonesplit-strength.png"
 
 export const phalanxbaneBlade = defineMartialArt({
   id: MARTIAL_ART_ID.phalanxbaneBlade,
