+++ src/data/martialArts/strategicSword.ts
@@ -1,6 +1,6 @@
 import { defineMartialArt } from "../../definitions/martialArts/martialArtDef"
 import { MARTIAL_ART_ID } from "./ids"
-import icon from "./icons/strategic-sword.png"
+import icon from "./icons/bellstrike-umbra.png"
 
 export const strategicSword = defineMartialArt({
   id: MARTIAL_ART_ID.strategicSword,
